STATISTICS-2: Migrate "o.a.c.math4.distribution" from Commons Math.
"IntegerDistribution" was renamed "DiscreteDistribution".
"RealDistribution" was renamed "ContinuousDistribution".

All exceptions are instances of "DistributionException" (package-private).

Solver code (used by method "inverseCumulativeProbability") is a private
static inner class in "AbstractContinuousDistribution".
Tolerances are hard-coded. [Constructors that specified a tolerance were
removed.]

Calls to "FastMath" were replaced by calls to JDK "Math".  This has led to
two unit tests failing in "GammaDistributionTest" for which the tolerance
had to be slightly increased. [The main source indicates which calls to
"Math" methods are responsible for the failures at the original tolerance.]