LogNormalDistribution: remove impossible condition
If x0 is not larger than x1, and x0 is larger than zero, then x1 cannot
be zero or below. Thus this condition:

if (x0 <= 0 || x1 <= 0)

can only be true in the second part if x0 is NaN. NaN checking is not
supported so we can drop the second condition.