Assert that DistributionException is thrown.
Asserting for IllegalArgumentException fails to detect if the formatting
done by DistributionException in the constructor is valid. If
String.format errors due to bad parameters it will throw an
IllegalArgumentException. This is an error but will not be detected in
the test suite (unless the returned exception is further asserted).