STATISTICS-32: New "survivalProbability" function for all discrete distributions.
While a naive implementation would simply be
`1-cumulativeProbability`, that would result
in loss of precision.

For many of the current discrete distributions a higher
precision survival probability is calculated.

For others, it is simply `1-cumulativeProbability`.

Many tests were added to verify the following:
- the precision of cumulativeProbability
- the precision of survivalProbability
- That survivalProbabiliy is near 1-cumulative
- That survival and cumulative probabilities are
  complementary

Through this development, certain distributions
were found lacking precision for their
cumulativeProbabilities and were improved.

These were:

- BinomialDistribution
- HypergeometricDistribuion

Expanding the tests for the Pascal distribution for the degenerate cases
found a bug in the p=1 degenerate case when x=0. The NaN return value
has been corrected to either 0 (x=0) or -infinity (x!=0).