Use faster and more robust RNG for tests.
Avoids the use of the WELL family of generators as they use a large
seed, are relatively slow and have less statistically robust output.

The RandomSource was selected using a generator with a period of
approximately 2^256 that passes PractRand for 4TiB of output. Fixed
seeds have been updated where appropriate.