Update base test class for Zipf distribution
Use power to the negative exponent in the generalized harmonic series
(to avoid the reciprocal with a divide).

Always sum the generalized harmonic from small values to large. Only
applies to the mean and variance computations.

Added survival function computation.

Allow exponent = 0.

Cache the log of the nth Harmonic for the logDensity computation.