STATISTICS-37: Update Normal distribution with high precision erfc
Commons Numbers 1.1 improved the implementation of erfc and its inverse.

Update the normal normal distribution and unit tests to verify the
increased
precision.

The increase in precision of the error function is such that the
computation of sd * sqrt(2) is a source of error against reference data.
This has been updated to compute exactly and verified against high
precision data. It is a small additional computation cost during class
initialisation.

Change extended precision CDF and SF data to use matlab variable
precision arithmetic.