STATISTICS-37: Update Normal distribution with high precision erfc
Commons Numbers 1.1 improved the implementation of erfc and its inverse.

Update the truncated normal distribution and unit tests to verify the
increased
precision.

Added better use of the probability in a range for the truncated normal.

Add notes about possible cancellation is computation of the moments for
the truncated normal.