STATISTIC-47: Add isf command to distribution examples application
Update check command to check the isf.

Change the check command to use Precision to test relative equality.

If the inverse probability does not map to the original input then do a
forward mapping to test the value maps back to the same probability.
This accommodates function pairs that are not bijections, i.e. are
many-to-one.