[STATISTICS-48] Remove isSupportConnected from distribution interface
This method is an implementation detail only used by the
AbstractContinuousDistribution to invert the cumulative or survival
probability. It has been demoted to a protected member of that class and
removed as a public interface member.

All distributions currently only return true for this method. Removal
from the public interface has no impact on the library.