Ensure default inverse probability is robust to distribution truncation
This was discovered as the F-distribution with numerator DF=1 and
denominator DF=2 computed a survival probability:

sf(infinity) = 0.0
sf(max_value) = 5.56e-309

If the inverse SF was called with a p-value lower than 5.56e-309 (e.g.
Double.MIN_VALUE) the required value of x is outside the finite range of
a double.