Update javadoc for mean and variance
Remove the statement 'or {@code Double.NaN} if it is not defined'.

This only applies to a few distributions. These distributions have been
updated to state they return NaN if the mean/variance is undefined.

This change prevents all distributions inheriting the javadoc and having
to override the return tag to state the mean/variance is finite.