Consistent javadoc for property getters
Use 'Gets the X parameter of this distribution'.

Removes the use of 'Access the X parameter' for some of the javadocs.
Using 'Gets' matches the same usage in the interface for the mean,
variance and lower/upper bounds.

Removes the use of parameter names within @code tags unless these are
defined in the same javadoc, e.g. LogNormalDistribution mu parameter.