Update f distribution pdf function
Greater accuracy is obtained using log(pdf) than using the log
computation. However the regular pdf may have intermediate overflow with
extreme parameters when the result is finite.

Extract the pdf and logpdf computation to a single method. Try the
regular pdf and switch to the log computation when the density is
sub-normal or overflows.