STATISTICS-70: Improve Hypergeometric distribution probability sums
Cache the midpoint of the CDF closest to 0.5. This midpoint is used to
compute the CDF or SF using the appropriate domain which uses the
summation of smaller terms.

The probability(x0, x1) function is implemented to sum the smallest
range of the PDF, avoiding the duplicate summation performed in the
default implementation using CDF(x1) - CDF(x0).

The inverse CDF or SF is computed using a single summation to find the
quantile.