STATISTICS-69: Add an unconditioned exact test for 2x2 tables
Support computing a test conditioned on the column sums using a
configurable method to identify more extreme tables. The p-value is
computed by maximising the probability over the more extreme tables.

The Commons Math BrentOptimizer has been extracted for use as a
univariate function optimizer. This has been modified to remove unused
functionality and support bounds on the function.