Replace the inference module DD class with Commons Numbers
Note: The DD class was ported to Commons Numbers in [NUMBERS-193].

Timings in the KolmogorovSmirnovDistributionTest from the previous
internal DD implementation have repeated on the same development machine
(but the JDK may have changed). The computed results using DD.pow or
DDMath.pow are an exact match. The new version is ~10% slower using
DD.pow; timings for DDMath.pow are approximately equal. The new times
have not been saved as these are only a reference guide.