Replace extended precision methods in distribution module with DD
All double-double methods can be implemented using the Numbers core DD
class.

Note on performance

xsqrt2pi: This removes a pre-split of the high part of a double-double
representation of sqrt(2pi) for multiplication. The split during the
method call to DD.multiply(double) uses 4 FLOPS.

sqrt2xx: The use of DD.sqrt also computes the round-off part adding 3
flops to the previous computation.

In both cases performance cost is negligible as the methods are invoked
only during class constructors.

expmhxx: The number of FLOPS for the method is unchanged.