STATISTICS-71: Add biased option for moment based statistics
Updates StandardDeviation, Variance, Skewness and Kurtosis.

Changes the computation of the skewness and kurtosis to detect zero
variance with a common method. The skewness now computes the unbiased G1
using a scaled biased g1. This provides a consistent result when
disabling the bias
correction by using the second moment to detect a zero variance
denominator.

When the variance is zero the result is returned as NaN (and not zero as
previously) due to 0 / 0. This matches the Scipy and Matlab result for
this case.