Update FirstMoment to use a half-representation
This maintains the overflow protection of downscaling but avoids
re-upscaling the moment and stored deviations for each input value.
Upscaling is only required when computing the final result.

This has a performance gain of 30-40%. Performance is approximately the
same as a rolling algorithm with no downscaling. Thus this modification
allows the overflow protection with negligible cost.

All sub-class moments must update their scaling factors when using the
deviations by appropriate powers of 2.